apps/predictions/ml_model.py
import numpy as np

# from sklearn.ensemble import RandomForestClassifier # Unused import
from xgboost import XGBClassifier
from datetime import datetime, timedelta
import os
import joblib  # Example for model persistence

# Define a path for saving/loading the model
MODEL_DIR = os.path.dirname(__file__)
MODEL_FILENAME = "wildfire_xgb_model.joblib"
MODEL_PATH = os.path.join(MODEL_DIR, MODEL_FILENAME)


class WildfirePredictionModel:
    def __init__(self):
        self.model = None
        self.version = "1.0.0"  # TODO: Link version to the loaded model file/metadata
        self.is_trained = False
        self._load_model()  # Attempt to load a pre-trained model on initialization

    def _load_model(self):
        """Loads the pre-trained XGBoost model from disk."""
        try:
            if os.path.exists(MODEL_PATH):
                self.model = joblib.load(MODEL_PATH)
                self.is_trained = True
                print(
                    f"Loaded pre-trained model version {self.version} from {MODEL_PATH}"
                )
            else:
                print(f"Model file not found at {MODEL_PATH}. Model needs training.")
                # Initialize a new model if none exists (optional, depends on workflow)
                self.model = XGBClassifier(
                    n_estimators=100,
                    max_depth=5,
                    learning_rate=0.1,
                    objective="binary:logistic",
                    # Add other relevant parameters
                )
                self.is_trained = False
        except Exception as e:
            print(f"Error loading model: {e}. Initializing a new model.")
            # Fallback to a new model instance on error
            self.model = XGBClassifier(
                n_estimators=100,
                max_depth=5,
                learning_rate=0.1,
                objective="binary:logistic",
            )
            self.is_trained = False

    def prepare_features(self, weather_data):
        """Prepares features from WeatherData. Assumes weather_data is a single instance."""
        # TODO: Feature engineering might be more complex.
        # TODO: Ensure consistency with features used during training.
        features = {
            "temperature": weather_data.temperature,
            "humidity": weather_data.humidity,
            "wind_speed": weather_data.wind_speed,
            "precipitation": weather_data.precipitation,
            "pressure": weather_data.pressure,
            "month": weather_data.timestamp.month,
            "day_of_year": weather_data.timestamp.timetuple().tm_yday,
            # Add other relevant features (e.g., vegetation index, drought index, historical fire data)
        }
        # Ensure the order of features matches the training data
        # This order MUST be consistent. Consider defining it explicitly.
        feature_order = [
            "temperature",
            "humidity",
            "wind_speed",
            "precipitation",
            "pressure",
            "month",
            "day_of_year",
        ]
        return {k: features[k] for k in feature_order}

    def train(self, X_train, y_train):
        """Trains the model. Requires a full data loading and preprocessing pipeline."""
        # TODO: Implement a full training pipeline:
        # 1. Load historical weather and wildfire data.
        # 2. Perform extensive feature engineering and selection.
        # 3. Preprocess data (scaling, encoding, imputation).
        # 4. Split data into train/validation/test sets.
        # 5. Train the model (potentially with hyperparameter tuning).
        # 6. Evaluate the model.
        # 7. Save the trained model.

        if self.model is None:
            print("Model not initialized. Cannot train.")
            return

        print("Starting model training...")  # Placeholder
        # Example: self.model.fit(X_train_processed, y_train)
        # self.is_trained = True
        # self._save_model() # Save after training
        print("Training requires a full implementation.")  # Placeholder
        pass

    def _save_model(self):
        """Saves the trained model to disk."""
        if self.model and self.is_trained:
            try:
                joblib.dump(self.model, MODEL_PATH)
                print(f"Model saved to {MODEL_PATH}")
            except Exception as e:
                print(f"Error saving model: {e}")
        else:
            print("Model is not trained or not initialized. Cannot save.")

    def predict(self, features):
        """Makes a prediction based on prepared features."""
        if not self.model or not self.is_trained:
            # Option 1: Raise error
            # raise ValueError("Model is not loaded or trained.")
            # Option 2: Return default/error value
            print("Model is not ready for predictions.")
            return None, None  # Or appropriate error indication

        # TODO: Apply the SAME feature scaling/preprocessing used during training.
        # Example: X = self.scaler.transform(np.array([list(features.values())]))
        try:
            X = np.array([list(features.values())])
            # Ensure X has the correct shape and feature order expected by the model

            risk_prob = self.model.predict_proba(X)[0][
                1
            ]  # Probability of class 1 (assume positive class is high risk)
            confidence = max(
                self.model.predict_proba(X)[0]
            )  # Probability of the predicted class

            return risk_prob, confidence
        except Exception as e:
            print(f"Error during prediction: {e}")
            return None, None  # Indicate prediction failure


# Example of how you might instantiate and use it elsewhere (e.g., in a view or management command)
# model_instance = WildfirePredictionModel()
# features = model_instance.prepare_features(some_weather_data_object)
# if features:
#    risk, conf = model_instance.predict(features)
