requirements.txt
# Core dependencies
Django==5.0.1
python-dotenv==1.0.0
requests==2.31.0
django-leaflet==0.29.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0

# API and Documentation
djangorestframework==3.14.0
drf-spectacular==0.27.1
markdown==3.5.2
django-filter==23.5

# Security
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0

# Data Science
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
xgboost==3.0.0
python-dateutil==2.8.2
plotly==6.0.1 